LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/pixel-manipulation/2d.imageData.put.unaffected-expected.txt

2d.imageData.put.unaffected
putImageData() is not affected by context state
Actual output:
Expected output:


PASS putImageData() is not affected by context state
